api/requirements.txt
Django==2.2
djangorestframework==3.12.4
mysqlclient==2.0.3
django-debug-toolbar==3.2.2
python-memcached==1.59